requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
# created-by: conda 24.9.2
antlr4-python3-runtime=4.9.3=pypi_0
anytree=2.13.0=pypi_0
bzip2=1.0.8=h99b78c6_7
ca-certificates=2025.6.15=hbd8a1cb_0
cmake=4.0.3=pypi_0
contourpy=1.3.0=pypi_0
cycler=0.12.1=pypi_0
derivative=0.6.3=pypi_0
fonttools=4.58.5=pypi_0
gsw=3.6.19=pypi_0
hydra-core=1.3.2=pypi_0
importlib-metadata=8.7.0=pypi_0
importlib-resources=6.5.2=pypi_0
joblib=1.5.1=pypi_0
kiwisolver=1.4.7=pypi_0
libexpat=2.7.0=h286801f_0
libffi=3.4.6=h1da3d7d_1
liblzma=5.8.1=h39f12f2_2
libsqlite=3.50.2=h6fb428d_0
libzlib=1.3.1=h8359307_2
matplotlib=3.9.4=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.5=h5e97a16_3
numpy=2.0.2=pypi_0
omegaconf=2.3.0=pypi_0
openssl=3.5.1=h81ee809_0
ordered-set=4.1.0=pypi_0
packaging=25.0=pypi_0
pandas=2.3.1=pypi_0
pillow=11.3.0=pypi_0
pip=25.1.1=pyh8b19718_0
pyparsing=3.2.3=pypi_0
pysindy=1.7.5=pypi_0
python=3.9.23=h7139b31_0_cpython
python-dateutil=2.9.0.post0=pypi_0
pytz=2025.2=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h1d1bf99_2
scikit-learn=1.6.1=pypi_0
scipy=1.13.1=pypi_0
scs=3.2.7.post2=pypi_0
seaborn=0.13.2=pypi_0
setuptools=80.9.0=pyhff2d567_0
six=1.17.0=pypi_0
sympy=1.14.0=pypi_0
threadpoolctl=3.6.0=pypi_0
tk=8.6.13=h892fb3f_2
tzdata=2025.2=pypi_0
wheel=0.45.1=pyhd8ed1ab_1
xarray=2024.7.0=pypi_0
zipp=3.23.0=pypi_0
